requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-allauth==0.57.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Cache & Sessions
redis==5.0.1
django-redis==5.4.0

# Web3 & Crypto
web3==6.11.3
eth-account==0.9.0
cryptography==41.0.7

# Background Tasks
celery==5.3.4
celery[redis]==5.3.4

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring & Logging
sentry-sdk==1.38.0
django-extensions==3.2.3

# Development & Testing
python-dotenv==1.0.0
factory-boy==3.3.0
coverage==7.3.2

# Security
django-ratelimit==4.1.0
django-csp==3.7

# Utilities
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3